move background code into its own file
